requirements.txt
Django==3.1.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
djangorestframework==3.12.1